Combine DefaultSolver implementations
