+++ examples/cpp/CMakeLists.txt
@@ -3,7 +3,6 @@ set(CPP_EXAMPLES
   example_expcone
   example_lp
   example_powcone
-  example_powcone_csc
   example_qp
   example_socp
   example_sdp
