+++ examples/cpp/example_powcone.cpp
@@ -62,7 +62,7 @@ int main()
                                            .build();
 
     // Build solver
-    clarabel::eigen::DefaultSolver<double> solver(P, q, A, b, cones, settings);
+    DefaultSolver<double> solver(P, q, A, b, cones, settings);
 
     // Solve
     solver.solve();
