+++ examples/cpp/example_qp.cpp
@@ -47,7 +47,7 @@ int main(void)
     DefaultSettings<double> settings = DefaultSettings<double>::default_settings();
 
     // Build solver
-    clarabel::eigen::DefaultSolver<double> solver(P, q, A, b, cones, settings);
+    DefaultSolver<double> solver(P, q, A, b, cones, settings);
 
     // Solve
     solver.solve();
