+++ examples/cpp/example_socp.cpp
@@ -45,7 +45,7 @@ int main()
     DefaultSettings<double> settings = DefaultSettings<double>::default_settings();
 
     // Build solver
-    clarabel::eigen::DefaultSolver<double> solver(P, q, A, b, cones, settings);
+    DefaultSolver<double> solver(P, q, A, b, cones, settings);
 
     // Solve
     solver.solve();
